fix some bjp3 issues. add getlightstyle builtins. fix some compile issues. fix decals on hlbsp.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5040 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/sound.h
@@ -331,8 +331,8 @@ struct soundcardinfo_s { //windows has one defined AFTER directsound
 //info on which sound effects are playing
 	//FIXME: use a linked list
 	channel_t	*channel;
-	int			total_chans;
-	int			max_chans;
+	size_t		total_chans;
+	size_t		max_chans;
 
 	float	ambientlevels[NUM_AMBIENTS];	//we use a float instead of the channel's int volume value to avoid framerate dependancies with slow transitions.
